206 updateDBPass doesn't display password on screen. Prompts to re-enter password to verify match.

+++ dsapp-test.sh
@@ -1906,42 +1906,52 @@ EOF
 }
 
 function changeDBPass {
-	local lineNumber input
+	local lineNumber input vinput
 	datasyncBanner;
-	read -p "Enter new database password: " input
-	if [ -z "$input" ];then
-		echo "Invalid input";
-		exit 1
+	if askYesOrNo "Change datasync database password?";then
+
+		read -sp "Enter new database password: " input
+		if [ -z "$input" ];then
+			echo "Invalid input";
+			exit 1
+		fi
+		echo
+		read -sp "Re-enter new database password: " vinput
+		if [ "$vinput" != "$input" ];then
+			echo "Passwords do not match"
+			exit 1
+		fi
+		echo
+		#Get Encrypted password from user input
+		local inputEncrpt=$(encodeString $input)
+
+		# Backup conf files
+		backupConf "changeDBPass";
+
+		echo "Changing database password"
+		su postgres -c "psql -c \"ALTER USER datasync_user WITH password '"$input"';\"" &>/dev/null
+		lineNumber=`cat --number $ceconf | sed -n "/<database>/,/<\/database>/p" | grep -i password | awk '{print $1}'`
+
+		if [[ $(isStringProtected /config/configengine/database $ceconf) -eq 1 ]];then
+			sed -i ""$lineNumber"s|<password>.*</password>|<password>"$inputEncrpt"</password>|g" $ceconf
+		else
+			sed -i ""$lineNumber"s|<password>.*</password>|<password>"$input"</password>|g" $ceconf
+		fi
+
+		if [[ $(isStringProtected /engine/settings/database $econf) -eq 1 ]];then
+			sed -i "s|<password>.*</password>|<password>"$inputEncrpt"</password>|g" $econf
+		else
+			sed -i "s|<password>.*</password>|<password>"$input"</password>|g" $econf
+		fi
+
+		if [[ $(isStringProtected /connector/settings/custom $mconf) -eq 1 ]];then
+			sed -i "s|<dbpass>.*</dbpass>|<dbpass>"$inputEncrpt"</dbpass>|g" $mconf
+		else
+			sed -i "s|<dbpass>.*</dbpass>|<dbpass>"$input"</dbpass>|g" $mconf
+		fi
+
+		echo -e "\nDatabase password updated. Please restart mobility."
 	fi
-	#Get Encrypted password from user input
-	local inputEncrpt=$(encodeString $input)
-
-	# Backup conf files
-	backupConf "changeDBPass";
-
-	echo "Changing database password"
-	su postgres -c "psql -c \"ALTER USER datasync_user WITH password '"$input"';\"" &>/dev/null
-	lineNumber=`cat --number $ceconf | sed -n "/<database>/,/<\/database>/p" | grep -i password | awk '{print $1}'`
-
-	if [[ $(isStringProtected /config/configengine/database $ceconf) -eq 1 ]];then
-		sed -i ""$lineNumber"s|<password>.*</password>|<password>"$inputEncrpt"</password>|g" $ceconf
-	else
-		sed -i ""$lineNumber"s|<password>.*</password>|<password>"$input"</password>|g" $ceconf
-	fi
-
-	if [[ $(isStringProtected /engine/settings/database $econf) -eq 1 ]];then
-		sed -i "s|<password>.*</password>|<password>"$inputEncrpt"</password>|g" $econf
-	else
-		sed -i "s|<password>.*</password>|<password>"$input"</password>|g" $econf
-	fi
-
-	if [[ $(isStringProtected /connector/settings/custom $mconf) -eq 1 ]];then
-		sed -i "s|<dbpass>.*</dbpass>|<dbpass>"$inputEncrpt"</dbpass>|g" $mconf
-	else
-		sed -i "s|<dbpass>.*</dbpass>|<dbpass>"$input"</dbpass>|g" $mconf
-	fi
-
-	echo -e "\nDatabase password updated. Please restart mobility."
 	eContinue;
 }
